break out FileItems component
